+++ src/main/java/br/com/digitaldevilsaga/dto/NovoBrinquedoDto.java
@@ -1,4 +1,6 @@
 package br.com.digitaldevilsaga.dto;
+import java.math.BigDecimal;
+
 import org.springframework.web.multipart.MultipartFile;
 
 public class NovoBrinquedoDto{
@@ -6,7 +8,7 @@ public class NovoBrinquedoDto{
     private String nome;
     private String descricao;
     private Integer idCategoria;
-    private double preco;
+    private BigDecimal preco;
     private MultipartFile imagem;
 
     public String getDescricao() {
@@ -32,11 +34,11 @@ public class NovoBrinquedoDto{
         this.idCategoria = idCategoria;
     }
 
-    public double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(double valor) {
+    public void setPreco(BigDecimal valor) {
         this.preco = valor;
     }
